fix help output, add snapshots
